Add m4a1 texture
Allow to shear textures to a fraction (float factors).

+++ src/game/Shearable.cpp
@@ -28,31 +28,31 @@ namespace PolandBall {
 
 namespace Game {
 
-void Shearable::shearX(int slice, int totalSlices) {
+void Shearable::shearX(float slice, int totalSlices) {
     if (slice >= totalSlices) {
         throw std::invalid_argument(Graphene::LogFormat("Slice value is invalid"));
     }
 
     this->shear.set(0, 0, 1.0f / totalSlices);
-    this->shear.set(0, 3, static_cast<float>(slice) / totalSlices);
+    this->shear.set(0, 3, slice / totalSlices);
 }
 
-void Shearable::shearY(int slice, int totalSlices) {
+void Shearable::shearY(float slice, int totalSlices) {
     if (slice >= totalSlices) {
         throw std::invalid_argument(Graphene::LogFormat("Slice value is invalid"));
     }
 
     this->shear.set(1, 1, 1.0f / totalSlices);
-    this->shear.set(1, 3, static_cast<float>(slice) / totalSlices);
+    this->shear.set(1, 3, slice / totalSlices);
 }
 
-void Shearable::shearZ(int slice, int totalSlices) {
+void Shearable::shearZ(float slice, int totalSlices) {
     if (slice >= totalSlices) {
         throw std::invalid_argument(Graphene::LogFormat("Slice value is invalid"));
     }
 
     this->shear.set(2, 2, 1.0f / totalSlices);
-    this->shear.set(2, 3, static_cast<float>(slice) / totalSlices);
+    this->shear.set(2, 3, slice / totalSlices);
 }
 
 const Math::Mat4& Shearable::getShear() const {
